Moved search result into separate component
